+++ include/pbc_bls.h
@@ -37,7 +37,7 @@ void pbc_sum_bytes_G2_compressed(element_s *elem_sum,
                                  size_t n,
                                  pairing_t pairing);
 void bls_sign_message(uint8_t *out_buf, element_s *sig_elem, element_s *hash_elem, uint8_t *msg,
-                      uint32_t msg_len, element_s *secret_key);
-int bls_verify_signature(element_s *sig, element_s *hash_elem, uint8_t *sig_buf, uint8_t *msg, uint32_t msg_len,
+                      uint64_t msg_len, element_s *secret_key);
+int bls_verify_signature(element_s *sig, element_s *hash_elem, uint8_t *sig_buf, uint8_t *msg, uint64_t msg_len,
                          element_s *public_key, element_s *g2, pairing_t pairing);
 #endif //ALPENHORN_PBC_SIGN_H
