+++ src/pbc_ibe.c
@@ -23,7 +23,7 @@ ibe_params *ibe_alloc(char *pb_params, const char *gen)
 	return params;
 }
 
-int ibe_pbc_extract(element_s *out, element_s *master_priv_key, const uint8_t *id, const uint32_t id_length)
+int ibe_pbc_extract(element_s *out, element_s *master_priv_key, const uint8_t *id, const uint64_t id_length)
 {
 	uint8_t id_hash[crypto_ghash_BYTES];
 	int res = crypto_generichash(id_hash, crypto_ghash_BYTES, id, id_length, NULL, 0);
@@ -52,7 +52,7 @@ void ibe_pbc_sk_from_hashes(uint8_t *sk_out,
 	crypto_generichash_final(&hash_state, sk_out, crypto_ghash_BYTES);
 }
 
-ssize_t ibe_pbc_encrypt(uint8_t *out, uint8_t *msg, uint32_t msg_len, element_s *public_key,
+ssize_t ibe_pbc_encrypt(uint8_t *out, uint8_t *msg, uint64_t msg_len, element_s *public_key,
                         element_s *gen, uint8_t *recv_id, size_t recv_id_len, pairing_s *pairing)
 {
 
@@ -100,7 +100,7 @@ ssize_t ibe_pbc_encrypt(uint8_t *out, uint8_t *msg, uint32_t msg_len, element_s 
 ssize_t
 ibe_pbc_decrypt(uint8_t *out,
                 uint8_t *c,
-                uint32_t clen,
+                uint64_t clen,
                 element_s *private_key,
                 uint8_t *public_key,
                 pairing_s *pairing)
